Move `encoding.go` to lib to avoid import cycle
