+++ app/encoding.go
@@ -6,7 +6,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/std"
 	"github.com/cosmos/cosmos-sdk/x/auth/tx"
 
-	"github.com/planetmint/planetmint-go/app/params"
+	"github.com/planetmint/planetmint-go/lib/params"
 )
 
 // makeEncodingConfig creates an EncodingConfig for an amino based test configuration.
